add simple rate limiting by configuration
Together with the configurable batch size it allows to specifically rate limit the throughput against the sink. This feature is mostly helpful in case the write load rate against the sink has to be throttled. As stated in the #62 description, the other kafka/connect/consumer configuration means to control this are sometimes not enough or hard to tweak. Rate limiting should most of the time not be needed at all which is why the default configuration uses settings which are "backwards-compatible" and do not apply any rate limits.